+++ CHANGELOG.md
@@ -15,11 +15,15 @@
   FunctionBodyLength, Nesting, TypeBodyLength, TypeName, VariableName.  
   [JP Simard](https://github.com/jpsim)
 
+* Add RuleEnabler, a struct that controls which rules are enabled and which are disabled.
+  [Aaron Daub](https://github.com/aarondaub)
+
 ##### Bug Fixes
 
 None.
 
 
 ## 0.1.0
+
 
 First Version!